Fold conversion-error tests into FormTests instead of a new file (#11828)
